refactor(party): simplified packages for business
