+++ mvp-gate-party/src/main/kotlin/cz/tyckouni/mvpgate/party/business/core/IdpService.kt
@@ -1,0 +1,5 @@
+package cz.tyckouni.mvpgate.party.business.core
+
+import cz.tyckouni.mvpgate.party.persistence.entity.Idp
+
+interface IdpService : Service<Idp>
