+++ mvp-gate-party/src/main/kotlin/cz/tyckouni/mvpgate/party/business/core/ServiceBase.kt
@@ -1,4 +1,4 @@
-package cz.tyckouni.mvpgate.party.business
+package cz.tyckouni.mvpgate.party.business.core
 
 import cz.tyckouni.mvpgate.party.persistence.entity.PartyEntity
 import cz.tyckouni.mvpgate.party.persistence.repository.PartyRepository
