+++ mvp-gate-party/src/main/kotlin/cz/tyckouni/mvpgate/party/business/core/SpService.kt
@@ -1,0 +1,13 @@
+package cz.tyckouni.mvpgate.party.business.core
+
+import cz.tyckouni.mvpgate.party.persistence.entity.Sp
+import java.util.UUID
+
+interface SpService : Service<Sp> {
+
+    fun findByName(name: String): Set<Sp>
+
+    fun create(name: String): Sp
+
+    fun delete(guid: UUID)
+}
