+++ mvp-gate-party/src/main/kotlin/cz/tyckouni/mvpgate/party/business/core/impl/IdpServiceImpl.kt
@@ -1,6 +1,7 @@
-package cz.tyckouni.mvpgate.party.business.idp
+package cz.tyckouni.mvpgate.party.business.core.impl
 
-import cz.tyckouni.mvpgate.party.business.ServiceBase
+import cz.tyckouni.mvpgate.party.business.core.ServiceBase
+import cz.tyckouni.mvpgate.party.business.core.IdpService
 import cz.tyckouni.mvpgate.party.persistence.entity.Idp
 import cz.tyckouni.mvpgate.party.persistence.repository.IdpRepository
 import org.springframework.stereotype.Service
