--- mvp-gate-party/src/main/kotlin/cz/tyckouni/mvpgate/party/business/idp/IdpService.kt
@@ -1,6 +1,0 @@
-package cz.tyckouni.mvpgate.party.business.idp
-
-import cz.tyckouni.mvpgate.party.business.Service
-import cz.tyckouni.mvpgate.party.persistence.entity.Idp
-
-interface IdpService : Service<Idp>
