+++ mvp-gate-party/src/main/kotlin/cz/tyckouni/mvpgate/party/grpc/IdpServiceGrpcImpl.kt
@@ -1,7 +1,7 @@
 package cz.tyckouni.mvpgate.party.grpc
 
 import cz.tyckouni.mvpgate.grpc.ByGuidRequest
-import cz.tyckouni.mvpgate.party.business.idp.IdpService
+import cz.tyckouni.mvpgate.party.business.core.IdpService
 import cz.tyckouni.mvpgate.party.grpc.converter.IdpGrpcConverter
 import io.micrometer.tracing.Tracer
 import org.lognet.springboot.grpc.GRpcService
