+++ mvp-gate-party/src/main/kotlin/cz/tyckouni/mvpgate/party/grpc/SpServiceGrpcImpl.kt
@@ -2,7 +2,7 @@ package cz.tyckouni.mvpgate.party.grpc
 
 import com.google.protobuf.Empty
 import cz.tyckouni.mvpgate.grpc.ByGuidRequest
-import cz.tyckouni.mvpgate.party.business.sp.SpService
+import cz.tyckouni.mvpgate.party.business.core.SpService
 import cz.tyckouni.mvpgate.party.grpc.converter.SpGrpcConverter
 import io.micrometer.tracing.Tracer
 import org.lognet.springboot.grpc.GRpcService
