+++ mvp-gate-party/src/main/kotlin/cz/tyckouni/mvpgate/party/persistence/repository/IdpRepository.kt
@@ -2,4 +2,7 @@ package cz.tyckouni.mvpgate.party.persistence.repository
 
 import cz.tyckouni.mvpgate.party.persistence.entity.Idp
 
+/**
+ * Persistence repository for the [Idp] JPA entity
+ */
 interface IdpRepository : PartyRepository<Idp>
