+++ mvp-gate-party/src/main/kotlin/cz/tyckouni/mvpgate/party/persistence/repository/SpRepository.kt
@@ -2,4 +2,7 @@ package cz.tyckouni.mvpgate.party.persistence.repository
 
 import cz.tyckouni.mvpgate.party.persistence.entity.Sp
 
+/**
+ * Persistence repository for the [Sp] JPA entity
+ */
 interface SpRepository : PartyRepository<Sp>
